+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.3.0-rc.0",
+  sdkVersion: "3.3.0-rc.1",
   genVersion: "2.558.5",
-  userAgent: "speakeasy-sdk/typescript 3.3.0-rc.0 2.558.5 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.3.0-rc.1 2.558.5 2.0 @orq-ai/node",
 } as const;