+++ packages/orq-rc/RELEASES.md
@@ -1029,3 +1029,13 @@ Based on:
 - [typescript v3.3.0-rc.0] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.0 - packages/orq-rc
+
+## 2025-03-21 10:09:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.521.1 (2.558.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.0-rc.1] packages/orq-rc
+### Releases
+- [NPM v3.3.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.1 - packages/orq-rc